+++ CHANGELOG.md
@@ -12,6 +12,7 @@ _Some previous changes will be added_
 ### Added
 
 * New exportable data : piece (UNIMARC 463, internal name is `linking_piece`) and document type
+* New action : query title, publisher and publication date in the _any_ index in Koha SRU (added to `SEARCH_IN_KOHA_SRU_VANILLA` operation)
 
 ### Changed